Add validation to the FDB version
