+++ api/v1beta1/foundationdbcluster_types.go
@@ -70,6 +70,7 @@ func init() {
 // FoundationDBClusterSpec defines the desired state of a cluster.
 type FoundationDBClusterSpec struct {
 	// Version defines the version of FoundationDB the cluster should run.
+	// +kubebuilder:validation:Pattern:=^(\d+)\.(\d+)\.(\d+)$
 	Version string `json:"version"`
 
 	// SidecarVersions defines the build version of the sidecar to run. This
@@ -2459,130 +2460,6 @@ type BuggifyConfig struct {
 	// CrashLoops defines a list of instance IDs that should be put into a
 	// crash looping state.
 	CrashLoop []string `json:"crashLoop,omitempty"`
-}
-
-// FdbVersion represents a version of FoundationDB.
-//
-// This provides convenience methods for checking features available in
-// different versions.
-type FdbVersion struct {
-	// Major is the major version
-	Major int
-
-	// Minor is the minor version
-	Minor int
-
-	// Patch is the patch version
-	Patch int
-}
-
-var fdbVersionRegex = regexp.MustCompile(`^(\d+)\.(\d+)\.(\d+)$`)
-
-// ParseFdbVersion parses a version from its string representation.
-func ParseFdbVersion(version string) (FdbVersion, error) {
-	matches := fdbVersionRegex.FindStringSubmatch(version)
-	if matches == nil {
-		return FdbVersion{}, fmt.Errorf("Could not parse FDB version from %s", version)
-	}
-
-	major, err := strconv.Atoi(matches[1])
-	if err != nil {
-		return FdbVersion{}, err
-	}
-
-	minor, err := strconv.Atoi(matches[2])
-	if err != nil {
-		return FdbVersion{}, err
-	}
-
-	patch, err := strconv.Atoi(matches[3])
-	if err != nil {
-		return FdbVersion{}, err
-	}
-
-	return FdbVersion{Major: major, Minor: minor, Patch: patch}, nil
-}
-
-// String gets the string representation of an FDB version.
-func (version FdbVersion) String() string {
-	return fmt.Sprintf("%d.%d.%d", version.Major, version.Minor, version.Patch)
-}
-
-// IsAtLeast determines if a version is greater than or equal to another version.
-func (version FdbVersion) IsAtLeast(other FdbVersion) bool {
-	if version.Major < other.Major {
-		return false
-	}
-	if version.Major > other.Major {
-		return true
-	}
-	if version.Minor < other.Minor {
-		return false
-	}
-	if version.Minor > other.Minor {
-		return true
-	}
-	if version.Patch < other.Patch {
-		return false
-	}
-	if version.Patch > other.Patch {
-		return true
-	}
-	return true
-}
-
-// IsProtocolCompatible determines whether two versions of FDB are protocol
-// compatible.
-func (version FdbVersion) IsProtocolCompatible(other FdbVersion) bool {
-	return version.Major == other.Major && version.Minor == other.Minor
-}
-
-// HasInstanceIDInSidecarSubstitutions determines if a version has
-// FDB_INSTANCE_ID supported natively in the variable substitutions in the
-// sidecar.
-func (version FdbVersion) HasInstanceIDInSidecarSubstitutions() bool {
-	return version.IsAtLeast(FdbVersion{Major: 6, Minor: 2, Patch: 15})
-}
-
-// PrefersCommandLineArgumentsInSidecar determines if a version has
-// support for configuring the sidecar exclusively through command-line
-// arguments.
-func (version FdbVersion) PrefersCommandLineArgumentsInSidecar() bool {
-	return version.IsAtLeast(FdbVersion{Major: 6, Minor: 2, Patch: 15})
-}
-
-// SupportsUsingBinariesFromMainContainer determines if a version has
-// support for having the sidecar dynamically switch between using binaries
-// from the main container and binaries provided by the sidecar.
-func (version FdbVersion) SupportsUsingBinariesFromMainContainer() bool {
-	return version.IsAtLeast(FdbVersion{Major: 6, Minor: 2, Patch: 15})
-}
-
-// HasRatekeeperRole determines if a version has a dedicated role for
-// ratekeeper.
-func (version FdbVersion) HasRatekeeperRole() bool {
-	return version.IsAtLeast(FdbVersion{Major: 6, Minor: 2, Patch: 0})
-}
-
-// HasMaxProtocolClientsInStatus determines if a version has the
-// max_protocol_clients field in the cluster status.
-func (version FdbVersion) HasMaxProtocolClientsInStatus() bool {
-	return version.IsAtLeast(FdbVersion{Major: 6, Minor: 2, Patch: 0})
-}
-
-// HasSidecarCrashOnEmpty determines if a version has the flag to have the
-// sidecar crash on a file being empty.
-func (version FdbVersion) HasSidecarCrashOnEmpty() bool {
-	return version.IsAtLeast(FdbVersion{Major: 6, Minor: 2, Patch: 20})
-}
-
-// HasNonBlockingExcludes determines if a version has support for non-blocking
-// exclude commands.
-//
-// This is currently set to false across the board, pending investigation into
-// potential bugs with non-blocking excludes.
-func (version FdbVersion) HasNonBlockingExcludes() bool {
-	return version.IsAtLeast(FdbVersion{Major: 6, Minor: 3, Patch: 5})
 }
 
 // PublicIPSource models options for how a pod gets its public IP.
