+++ config/crd/bases/apps.foundationdb.org_foundationdbclusters.yaml
@@ -7600,6 +7600,7 @@ spec:
                 updatePodsByReplacement:
                   type: boolean
                 version:
+                  pattern: ^(\d+)\.(\d+)\.(\d+)$
                   type: string
                 volumeClaim:
                   properties:
